misc/buy_buy_buy/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
discord-py-slash-command==2.3.2
discord.py==1.7.1
dnspython==2.1.0
idna==2.10
multidict==5.1.0
pymongo==3.11.3
python-dateutil==2.8.1
python-dotenv==0.19.0
requests==2.25.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.4
yarl==1.6.3
